refactor tweet mappers
